FutureYou/adminWebsite/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv
requests
pandas
openpyxl
python-docx
gunicorn
